delete comment when changing numbers protection

+++ delete_comment.php
@@ -21,26 +21,41 @@ if(isset($_POST['delete_comment_btn']) && !empty($_POST['comment_id']) && !empty
 		exit;
 	}
 
-    //check if Session id is the same as the owner.
+    $user_id = $_POST['user_id'];
+	$session_id = $_SESSION['id'];
+    try{
+		$conn = connect_PDO();
+		$stmt = $conn->prepare("SELECT * FROM comments WHERE user_id = ? AND id = ?");
+		$stmt->bindParam(1, $session_id, PDO::PARAM_INT);
+		$stmt->bindParam(2, $comment_id, PDO::PARAM_INT);
+		$stmt->execute();
+		$comment_from_db = $stmt->fetch(PDO::FETCH_ASSOC);
+	} catch (PDOException $error) {
+		echo $error->getMessage(); 
+		exit;
+	}
 
-    
+	if (isset($_POST['delete_comment_btn']) && $comment_id == $comment_from_db['id'] && $_SESSION['id'] == $user_id){
+        try {
+            $conn = connect_PDO();
+            $stmt = $conn->prepare("DELETE FROM comments WHERE id = ?");
+            $stmt->bindParam(1, $comment_id, PDO::PARAM_INT);
 
-    try {
-    $conn = connect_PDO();
-    $stmt = $conn->prepare("DELETE FROM comments WHERE id = ?");
-    $stmt->bindParam(1, $comment_id, PDO::PARAM_INT);
+            if($stmt->execute()){
+                header("location: single_post.php?post_id=".$post_id."&success_message=Comment deleted successfully");
+            }else{
+                header("location: single_post.php?post_id=".$post_id."&error_message=Could not delete comment");
+            }
 
-    if($stmt->execute()){
-        header("location: single_post.php?post_id=".$post_id."&success_message=Comment deleted successfully");
-    }else{
-        header("location: single_post.php?post_id=".$post_id."&error_message=Could not delete comment");
-    }
-
-    exit;
-    }
-    catch (PDOException $e) {
-            echo $e->getMessage();
-    }
+            exit;
+            }
+            catch (PDOException $e) {
+                    echo $e->getMessage();
+            }
+    } else {
+		header('location: index.php?error_message=You are going way too far in the tests for the evaluation!! ;)');
+		exit;
+	}
 
 }else{
     header("location: index.php?error_message=error");
